% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{bot_server}
\alias{bot_server}
\title{ZENIT-POLAR substitution cipher method of encryption}
\usage{
bot_server(token_path)
}
\arguments{
\item{token_path}{second one-way cipher}
}
\value{
encrypted message
}
\description{
Returns a message encrypted by ZENIT-POLAR substitution cipher method of encryption using by default TENIS-POLAR key.
}
\details{
This kind of encryption replace the nth character of 1th key to the nth character of 2nd key and vice versa.
}
