% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{generate_token}
\alias{generate_token}
\title{ZENIT-POLAR substitution cipher method of encryption}
\usage{
generate_token(
  json_path = "artifacts/keys_tokens.json",
  app_name = "Holiday-bot",
  token_path = "artifacts/rtweet_token.rds"
)
}
\arguments{
\item{json_path}{message to encrypt}

\item{app_name}{first one-way cipher}

\item{token_path}{second one-way cipher}
}
\value{
encrypted message
}
\description{
Returns a message encrypted by ZENIT-POLAR substitution cipher method of encryption using by default TENIS-POLAR key.
}
\details{
This kind of encryption replace the nth character of 1th key to the nth character of 2nd key and vice versa.
}
