% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{bot_server}
\alias{bot_server}
\title{Twitter bot server main function}
\usage{
bot_server(settings_path)
}
\arguments{
\item{settings_path}{second one-way cipher}
}
\value{
encrypted message
}
\description{
Posts a tweet based on ANBIMA holiday database
}
\details{
Given twitter bot token it tweets a message concerns holidays
}
