% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{generate_token}
\alias{generate_token}
\title{Token Generation}
\usage{
generate_token(settings_path)
}
\arguments{
\item{settings_path}{path and filename of settings file}
}
\value{
rds object
}
\description{
Creates a twitter bot token given twitter api credentials
}
\details{
Saves a twitter bot token as rds object
}
